+++ tests/backends/matplotlib/canvas_test.py
