+++ tests/backends/matplotlib/image_test.py
