+++ tests/backends/matplotlib/interactive_test.py
