--- tests/graphics/fig2d_test.py
@@ -1,259 +1,0 @@
-# SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
-
-from plopp.data.testing import data_array
-from plopp.graphics.fig2d import Figure2d
-from plopp.graphics.mesh import Mesh
-from plopp import input_node
-import matplotlib.pyplot as plt
-import numpy as np
-import scipp as sc
-import pytest
-import tempfile
-import os
-
-
-def test_empty():
-    fig = Figure2d()
-    assert len(fig.artists) == 0
-
-
-def test_update():
-    fig = Figure2d()
-    assert len(fig.artists) == 0
-    da = data_array(ndim=2)
-    key = 'data2d'
-    fig.update(da, key=key)
-    assert isinstance(fig.artists[key], Mesh)
-    assert sc.identical(fig.artists[key]._data, da)
-
-
-def test_update_not_2d_raises():
-    fig = Figure2d()
-    with pytest.raises(ValueError, match="Figure2d can only be used to plot 2-D data."):
-        fig.update(data_array(ndim=1), key='data1d')
-    with pytest.raises(ValueError, match="Figure2d can only be used to plot 2-D data."):
-        fig.update(data_array(ndim=3), key='data3d')
-
-
-def test_create_with_node():
-    da = data_array(ndim=2)
-    fig = Figure2d(input_node(da))
-    assert len(fig.artists) == 1
-    assert sc.identical(list(fig.artists.values())[0]._data, da)
-
-
-def test_create_with_bin_edges():
-    da = data_array(ndim=2, binedges=True)
-    fig = Figure2d(input_node(da))
-    assert len(fig.artists) == 1
-    assert sc.identical(list(fig.artists.values())[0]._data, da)
-
-
-def test_create_with_only_one_bin_edge_coord():
-    da = data_array(ndim=2, binedges=True)
-    da.coords['xx'] = sc.midpoints(da.coords['xx'])
-    fig = Figure2d(input_node(da))
-    assert len(fig.artists) == 1
-    assert sc.identical(list(fig.artists.values())[0]._data, da)
-
-
-def test_log_norm():
-    fig = Figure2d(norm='log')
-    assert fig.colormapper.norm == 'log'
-
-
-def test_toggle_norm():
-    da = data_array(ndim=2)
-    fig = Figure2d(input_node(da))
-    assert fig.colormapper.norm == 'linear'
-    fig.toggle_norm()
-    assert fig.colormapper.norm == 'log'
-
-
-def test_crop():
-    da = data_array(ndim=2, binedges=True)
-    fig = Figure2d(input_node(da))
-    assert fig.canvas.ax.get_xlim() == (da.meta['xx'].min().value,
-                                        da.meta['xx'].max().value)
-    assert fig.canvas.ax.get_ylim() == (da.meta['yy'].min().value,
-                                        da.meta['yy'].max().value)
-    xmin = sc.scalar(2.1, unit='m')
-    xmax = sc.scalar(102.0, unit='m')
-    ymin = sc.scalar(5.5, unit='m')
-    ymax = sc.scalar(22.3, unit='m')
-    fig.crop(xx={'min': xmin, 'max': xmax}, yy={'min': ymin, 'max': ymax})
-    assert fig.canvas.ax.get_xlim() == (xmin.value, xmax.value)
-    assert fig.canvas.ax.get_ylim() == (ymin.value, ymax.value)
-
-
-def test_crop_no_variable():
-    da = data_array(ndim=2)
-    xmin = 2.1
-    xmax = 102.0
-    ymin = 5.5
-    ymax = 22.3
-    fig = Figure2d(input_node(da),
-                   crop={
-                       'xx': {
-                           'min': xmin,
-                           'max': xmax
-                       },
-                       'yy': {
-                           'min': ymin,
-                           'max': ymax
-                       }
-                   })
-    assert fig.canvas.ax.get_xlim() == (xmin, xmax)
-    assert fig.canvas.ax.get_ylim() == (ymin, ymax)
-
-
-def test_cbar():
-    da = data_array(ndim=2, binedges=True)
-    fig = Figure2d(input_node(da), cbar=False)
-    assert fig.canvas.cax is None
-
-
-def test_update_on_one_mesh_changes_colors_on_second_mesh():
-    da1 = data_array(ndim=2)
-    da2 = 3.0 * data_array(ndim=2)
-    da2.coords['xx'] += sc.scalar(50.0, unit='m')
-    a = input_node(da1)
-    b = input_node(da2)
-    f = Figure2d(a, b)
-    old_b_colors = f.artists[b.id]._mesh.get_facecolors()
-    a.func = lambda: da1 * 1.1
-    a.notify_children('updated a')
-    # No change because the update did not change the colorbar limits
-    assert np.allclose(old_b_colors, f.artists[b.id]._mesh.get_facecolors())
-    a.func = lambda: da1 * 5.0
-    a.notify_children('updated a')
-    assert not np.allclose(old_b_colors, f.artists[b.id]._mesh.get_facecolors())
-
-
-def test_with_string_coord():
-    strings = ['a', 'b', 'c', 'd', 'e']
-    da = sc.DataArray(data=sc.array(dims=['y', 'x'], values=np.random.random((5, 5))),
-                      coords={
-                          'x': sc.array(dims=['x'], values=strings, unit='s'),
-                          'y': sc.arange('y', 5., unit='m')
-                      })
-    fig = Figure2d(input_node(da))
-    assert [t.get_text() for t in fig.canvas.ax.get_xticklabels()] == strings
-
-
-def test_with_strings_as_bin_edges():
-    strings = ['a', 'b', 'c', 'd', 'e', 'f']
-    da = sc.DataArray(data=sc.array(dims=['y', 'x'], values=np.random.random((5, 5))),
-                      coords={
-                          'x': sc.array(dims=['x'], values=strings, unit='s'),
-                          'y': sc.arange('y', 6., unit='m')
-                      })
-    fig = Figure2d(input_node(da))
-    assert [t.get_text() for t in fig.canvas.ax.get_xticklabels()] == strings
-
-
-def test_with_strings_as_bin_edges_other_coord_is_bin_centers():
-    strings = ['a', 'b', 'c', 'd', 'e', 'f']
-    da = sc.DataArray(data=sc.array(dims=['y', 'x'], values=np.random.random((5, 5))),
-                      coords={
-                          'x': sc.array(dims=['x'], values=strings, unit='s'),
-                          'y': sc.arange('y', 5., unit='m')
-                      })
-    fig = Figure2d(input_node(da))
-    assert [t.get_text() for t in fig.canvas.ax.get_xticklabels()] == strings
-
-
-def test_kwargs_are_forwarded_to_artist():
-    da = data_array(ndim=2)
-    fig = Figure2d(input_node(da), rasterized=True)
-    artist = list(fig.artists.values())[0]
-    assert artist._mesh.get_rasterized()
-    fig = Figure2d(input_node(da), rasterized=False)
-    artist = list(fig.artists.values())[0]
-    assert not artist._mesh.get_rasterized()
-
-
-def test_figsize():
-    da = data_array(ndim=2)
-    size = (8.1, 8.3)
-    fig = Figure2d(input_node(da), figsize=size)
-    assert np.allclose(fig.canvas.fig.get_size_inches(), size)
-
-
-def test_grid():
-    da = data_array(ndim=2)
-    fig = Figure2d(input_node(da), grid=True)
-    assert fig.canvas.ax.xaxis.get_gridlines()[0].get_visible()
-
-
-def test_ax():
-    fig, ax = plt.subplots()
-    assert len(ax.collections) == 0
-    da = data_array(ndim=2)
-    _ = Figure2d(input_node(da), ax=ax)
-    assert len(ax.collections) == 1
-
-
-def test_cax():
-    fig, ax = plt.subplots()
-    cax = fig.add_axes([0.9, 0.02, 0.05, 0.98])
-    assert len(cax.collections) == 0
-    da = data_array(ndim=2)
-    _ = Figure2d(input_node(da), ax=ax, cax=cax)
-    assert len(cax.collections) > 0
-
-
-@pytest.mark.parametrize('ext', ['jpg', 'png', 'pdf', 'svg'])
-def test_save_to_disk(ext):
-    da = data_array(ndim=2)
-    fig = Figure2d(input_node(da))
-    with tempfile.TemporaryDirectory() as path:
-        fname = os.path.join(path, f'plopp_fig2d.{ext}')
-        fig.save(filename=fname)
-        assert os.path.isfile(fname)
-
-
-def test_raises_for_new_data_with_incompatible_dimension():
-    a = data_array(ndim=2)
-    b = a.rename(xx='zz')
-    with pytest.raises(sc.DimensionError):
-        Figure2d(input_node(a), input_node(b))
-
-
-def test_raises_for_new_data_with_incompatible_unit():
-    a = data_array(ndim=2)
-    b = a * a
-    with pytest.raises(sc.UnitError):
-        Figure2d(input_node(a), input_node(b))
-
-
-def test_raises_for_new_data_with_incompatible_coord_unit():
-    a = data_array(ndim=2)
-    b = a.copy()
-    b.coords['xx'] = a.coords['xx'] * a.coords['xx']
-    with pytest.raises(sc.UnitError):
-        Figure2d(input_node(a), input_node(b))
-
-
-def test_converts_new_data_units():
-    a = data_array(ndim=2, unit='m')
-    b = data_array(ndim=2, unit='cm')
-    anode = input_node(a)
-    bnode = input_node(b)
-    fig = Figure2d(anode, bnode)
-    assert sc.identical(fig.artists[anode.id]._data, a)
-    assert sc.identical(fig.artists[bnode.id]._data, b.to(unit='m'))
-
-
-def test_converts_new_data_coordinate_units():
-    a = data_array(ndim=2)
-    b = data_array(ndim=2)
-    b.coords['xx'].unit = 'cm'
-    anode = input_node(a)
-    bnode = input_node(b)
-    fig = Figure2d(anode, bnode)
-    assert sc.identical(fig.artists[anode.id]._data, a)
-    c = b.copy()
-    c.coords['xx'] = c.coords['xx'].to(unit='m')
-    assert sc.identical(fig.artists[bnode.id]._data, c)
